requirements.txt
# Data Processing
pyspark==3.5.0
kafka-python==2.0.2
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data Visualization
matplotlib==3.8.2
plotly==5.18.0
streamlit==1.29.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Utilities
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2

# Documentation
mkdocs==1.5.3
mkdocs-material==9.4.7 